+++ src/App.tsx
@@ -4,6 +4,7 @@ import { InstallPromptRoot } from './components/InstallPromptRoot'
 import { InstallReturnVisitSync } from './components/install/InstallReturnVisitSync'
 import { AppShell } from './components/layout/AppShell'
 import { TrackerProvider } from './context/TrackerProvider'
+import { GitHubSyncProvider } from './context/GitHubSyncProvider'
 import { ElaraProvider } from './context/ElaraProvider'
 import { ToastProvider } from './context/ToastProvider'
 import type { TabId } from './types'
@@ -16,10 +17,12 @@ function App() {
       <ToastProvider>
         <InstallPromptRoot>
           <TrackerProvider>
-            <InstallReturnVisitSync />
-            <ElaraProvider>
-              <AppShell activeTab={activeTab} onTabChange={setActiveTab} />
-            </ElaraProvider>
+            <GitHubSyncProvider>
+              <InstallReturnVisitSync />
+              <ElaraProvider>
+                <AppShell activeTab={activeTab} onTabChange={setActiveTab} />
+              </ElaraProvider>
+            </GitHubSyncProvider>
           </TrackerProvider>
         </InstallPromptRoot>
       </ToastProvider>